+++ src/Products/PloneMeeting/tests/testAdvices.py
@@ -4242,6 +4242,15 @@ class testAdvices(PloneMeetingTestCase):
         self.assertTrue(self.hasPermission(View, advice))
         self.assertTrue(contenthistory.show_history())
 
+    def test_pm_get_advice_given_by(self):
+        """Show info "Given by" on advice."""
+        item, advice = self._setupItemWithAdvice()
+        view = item.restrictedTraverse('@@advice-infos')
+        view(advice.advice_group, False, item.adapted().getCustomAdviceMessageFor(advice))
+        # with default advice workflow, we do not manage advice_given_by
+        # as we only know who created the advice
+        self.assertIsNone(view.get_advice_given_by())
+
 
 def test_suite():
     from unittest import makeSuite